requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via litellm
aiosignal==1.4.0
    # via aiohttp
alembic==1.17.2
    # via optuna
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   asyncer
    #   dspy
    #   httpx
    #   openai
asyncer==0.0.8
    # via dspy
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via dspy
cachetools==6.2.2
    # via dspy
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   rac-dspy (pyproject.toml)
    #   litellm
    #   typer-slim
cloudpickle==3.1.2
    # via dspy
colorama==0.4.6
    # via
    #   click
    #   colorlog
    #   tqdm
colorlog==6.10.1
    # via optuna
diskcache==5.6.3
    # via dspy
distro==1.9.0
    # via openai
dspy==3.0.4
    # via dspy-ai
dspy-ai==3.0.4
    # via rac-dspy (pyproject.toml)
fastuuid==0.14.0
    # via litellm
filelock==3.20.0
    # via huggingface-hub
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.10.0
    # via huggingface-hub
gepa==0.0.17
    # via dspy
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   huggingface-hub
    #   langgraph-sdk
    #   langsmith
    #   litellm
    #   openai
huggingface-hub==1.1.5
    # via tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via litellm
jiter==0.12.0
    # via openai
joblib==1.5.2
    # via
    #   dspy
    #   scikit-learn
json-repair==0.54.2
    # via dspy
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.1
    # via litellm
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain-core==1.1.0
    # via
    #   rac-dspy (pyproject.toml)
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langgraph==1.0.4
    # via rac-dspy (pyproject.toml)
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.2.10
    # via langgraph
langsmith==0.4.48
    # via langchain-core
litellm==1.80.5
    # via dspy
magicattr==0.1.6
    # via dspy
mako==1.3.10
    # via alembic
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
numpy==2.3.5
    # via
    #   rac-dspy (pyproject.toml)
    #   dspy
    #   optuna
    #   pandas
    #   rank-bm25
    #   scikit-learn
    #   scipy
openai==2.8.1
    # via
    #   dspy
    #   litellm
optuna==4.6.0
    # via dspy
orjson==3.11.4
    # via
    #   dspy
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   optuna
pandas==2.3.3
    # via rac-dspy (pyproject.toml)
pillow==12.0.0
    # via dspy
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pydantic==2.12.4
    # via
    #   rac-dspy (pyproject.toml)
    #   dspy
    #   langchain-core
    #   langgraph
    #   langsmith
    #   litellm
    #   openai
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via litellm
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   langchain-core
    #   optuna
rank-bm25==0.2.2
    # via rac-dspy (pyproject.toml)
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.11.3
    # via
    #   dspy
    #   tiktoken
requests==2.32.5
    # via
    #   dspy
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==14.2.0
    # via
    #   rac-dspy (pyproject.toml)
    #   dspy
rpds-py==0.29.0
    # via
    #   jsonschema
    #   referencing
scikit-learn==1.7.2
    # via rac-dspy (pyproject.toml)
scipy==1.16.3
    # via scikit-learn
shellingham==1.5.4
    # via huggingface-hub
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.44
    # via
    #   alembic
    #   optuna
tenacity==9.1.2
    # via
    #   dspy
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.12.0
    # via litellm
tokenizers==0.22.1
    # via litellm
tqdm==4.67.1
    # via
    #   dspy
    #   huggingface-hub
    #   openai
    #   optuna
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   alembic
    #   anyio
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
xxhash==3.6.0
    # via
    #   dspy
    #   langgraph
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
zstandard==0.25.0
    # via langsmith
